Liz Comments on Homepage text
We need to remove the "fly in" of the apps next to the main text. It is distracting. Rather, we need to have 3-4 buttons that group all the solvers/apps that we will use together into classifications: Evaluate (Control Chart); Explore (Kriging); Allocate (NMF); Predict (?). 

Excited to actually run through these apps with you guys and start changing the text in the other sections!